Fix: user statistics on empty accounts
- Fix an issue where API returned error when trying to get statistics from an empty account
- Add a custom error when trying to delete an exercise that still has logs associated with it
- Stardardize token issuer

+++ src/main/java/dev/matheusmisumoto/workoutloggerapi/repository/UserRepository.java
@@ -17,7 +17,7 @@ public interface UserRepository extends JpaRepository<User, UUID> {
 	int totalWorkouts(UUID userId);
 	
 	@Query(value = "SELECT SUM(s.weight * s.reps) FROM workoutlogger.workouts_sets s INNER JOIN workoutlogger.workouts w ON w.id = s.workout_id WHERE w.user_id = ?1 GROUP BY w.user_id", nativeQuery = true)
-	int calculateUserTotalWeightLifted(UUID userId);
+	Double calculateUserTotalWeightLifted(UUID userId);
 	
 	UserDetails findByLogin(String login);
 }
